+++ src/compas_rcf/abb/__init__.py
@@ -11,3 +11,6 @@ Fabrication communication is setup using ROS and compas_rrc with the controller.
 This package should be used for any helpers or utilities we might require.
 
 """
+
+from .helpers import *  # noqa: F401,F403
+from .programs import *  # noqa: F401,F403