+++ src/compas_rcf/fabrication/abb_runner.py
@@ -7,316 +7,40 @@ from __future__ import division
 from __future__ import print_function
 
 import json
-import logging
+import logging as log
 import sys
 import time
 from datetime import datetime
 from operator import attrgetter
-from pathlib import Path
+import pathlib
 
 from compas.geometry import Frame
 from compas.geometry import Point
 from compas.geometry import Vector
 from compas_fab.backends.ros import RosClient
 from compas_rrc import AbbClient
-from compas_rrc import CustomInstruction
-from compas_rrc import MoveToFrame
-from compas_rrc import MoveToJoints
 from compas_rrc import PrintText
-from compas_rrc import ReadWatch
-from compas_rrc import SetAcceleration
-from compas_rrc import SetDigital
-from compas_rrc import SetMaxSpeed
-from compas_rrc import SetTool
-from compas_rrc import SetWorkObject
-from compas_rrc import StartWatch
-from compas_rrc import StopWatch
-from compas_rrc import WaitTime
 
 from compas_rcf import __version__
-from compas_rcf.abb.helpers import docker_compose_paths
-from compas_rcf.abb.helpers import ping
-from compas_rcf.abb.helpers import robot_ips
+from compas_rcf.abb import connection_check
+from compas_rcf.abb import docker_compose_paths
+from compas_rcf.abb import robot_ips
+from compas_rcf.abb import pick_bullet
+from compas_rcf.abb import place_bullet
+from compas_rcf.abb import post_procedure
+from compas_rcf.abb import pre_procedure
+from compas_rcf.docker import compose_up
 from compas_rcf.fabrication.clay_obj import ClayBulletEncoder
-from compas_rcf.fabrication.conf import abb_rcf_conf_template
-from compas_rcf.fabrication.conf import fabrication_conf
+from compas_rcf.fabrication.conf import FABRICATION_CONF as fab_conf
+from compas_rcf.fabrication.conf import interactive_conf_setup
+from compas_rcf.fabrication.conf import Path
 from compas_rcf.utils import ui
-from compas_rcf.utils.docker import compose_up
 from compas_rcf.utils.json_ import load_bullets
-from compas_rcf.utils.util_funcs import get_offset_frame
 
 if sys.version_info[0] < 2:
     raise Exception("This module requires Python 3")
 else:
     import questionary
-
-################################################################################
-# Globals                                                                      #
-################################################################################
-
-ROBOT_CONTROL_FOLDER_DRIVE = Path(
-    "G:\\Shared drives\\2020_MAS\\T2_P1\\02_Groups\\Phase2\\rcf_fabrication\\02_robot_control"  # noqa E501
-)
-DEFAULT_CONF_DIR = ROBOT_CONTROL_FOLDER_DRIVE / "05_fabrication_confs"
-DEFAULT_JSON_DIR = ROBOT_CONTROL_FOLDER_DRIVE / "04_fabrication_data_jsons"
-DEFAULT_LOG_DIR = ROBOT_CONTROL_FOLDER_DRIVE / "06_fabrication_logs"
-
-# Define external axis, will not be used but required in move cmds
-EXTERNAL_AXIS_DUMMY: list = []
-
-################################################################################
-# Programs                                                                     #
-################################################################################
-
-
-def initial_setup(client):
-    """Pre fabrication setup, speed, acceleration, tool, work object and initial pose.
-
-    Parameters
-    ----------
-    client : :class:`compas_rrc.AbbClient`
-    """
-    send_grip_release(client, CONF.tool.release_state)
-
-    client.send(SetTool(CONF.tool.tool_name))
-    logging.debug("Tool {} set.".format(CONF.tool.tool_name))
-    client.send(SetWorkObject(CONF.wobjs.placing_wobj_name))
-    logging.debug("Work object {} set.".format(CONF.wobjs.placing_wobj_name))
-
-    # Set Acceleration
-    client.send(SetAcceleration(CONF.speed_values.accel, CONF.speed_values.accel_ramp))
-    logging.debug("Acceleration values set.")
-
-    # Set Max Speed
-    client.send(
-        SetMaxSpeed(CONF.speed_values.speed_override, CONF.speed_values.speed_max_tcp)
-    )
-    logging.debug("Speed set.")
-
-    # Initial configuration
-    client.send(
-        MoveToJoints(
-            CONF.safe_joint_positions.start,
-            EXTERNAL_AXIS_DUMMY,
-            CONF.movement.speed_travel,
-            CONF.movement.zone_travel,
-        )
-    )
-    logging.debug("Sent move to safe joint position")
-
-
-def shutdown_procedure(client):
-    """Post fabrication procedure, end pose and closing and termination of client.
-
-    Parameters
-    ----------
-    client : :class:`compas_rrc.AbbClient`
-    """
-    send_grip_release(client, CONF.tool.release_state)
-
-    client.send(
-        MoveToJoints(
-            CONF.safe_joint_positions.end,
-            EXTERNAL_AXIS_DUMMY,
-            CONF.movement.speed_travel,
-            CONF.movement.zone_travel,
-        )
-    )
-
-    client.send_and_wait(PrintText("Finished"))
-
-    # Close client
-    client.close()
-    client.terminate()
-
-
-def send_picking(client, picking_frame):
-    """Send movement and IO instructions to pick up a clay bullet.
-
-    Parameters
-    ----------
-    client : :class:`compas_rrc.AbbClient`
-    picking_frame : compas.geometry.Frame
-        Target frame to pick up bullet
-    """
-    if CONF.target == "virtual":
-        # Custom instruction create a clay bullet in RobotStudio
-        # TODO: Create bullet at picking point
-        client.send(CustomInstruction("r_A057_RS_Create_Bullet"))
-
-    # change work object before picking
-    client.send(SetWorkObject(CONF.wobjs.picking_wobj_name))
-
-    # pick bullet
-    offset_picking = get_offset_frame(picking_frame, CONF.movement.offset_distance)
-
-    # start watch
-    client.send(StartWatch())
-
-    client.send(
-        MoveToFrame(
-            offset_picking, CONF.movement.speed_travel, CONF.movement.zone_travel
-        )
-    )
-
-    client.send(
-        MoveToFrame(
-            picking_frame, CONF.movement.speed_travel, CONF.movement.zone_travel
-        )
-    )
-
-    send_grip_release(client, CONF.tool.grip_state)
-
-    client.send(
-        MoveToFrame(
-            offset_picking, CONF.movement.speed_picking, CONF.movement.zone_pick
-        )
-    )
-
-    client.send(StopWatch())
-    return client.send(ReadWatch())
-
-
-def send_placing(client, bullet):
-    """Send movement and IO instructions to place a clay bullet.
-
-    Parameters
-    ----------
-    client : :class:`compas_rrc.AbbClient`
-    picking_frame : compas.geometry.Frame
-        Target frame to pick up bullet
-    """
-    logging.debug("Location frame: {}".format(bullet.location))
-
-    # change work object before placing
-    client.send(SetWorkObject(CONF.wobjs.placing_wobj_name))
-
-    # add offset placing plane to pre and post frames
-
-    top_bullet_frame = get_offset_frame(bullet.location, bullet.height)
-    offset_placement = get_offset_frame(top_bullet_frame, CONF.movement.offset_distance)
-
-    # start watch
-    client.send(StartWatch())
-
-    # Safe pos then vertical offset
-    for frame in bullet.trajectory_to:
-        client.send(
-            MoveToFrame(frame, CONF.movement.speed_travel, CONF.movement.zone_travel)
-        )
-
-    client.send(
-        MoveToFrame(
-            offset_placement, CONF.movement.speed_travel, CONF.movement.zone_travel
-        )
-    )
-    client.send(
-        MoveToFrame(
-            top_bullet_frame, CONF.movement.speed_placing, CONF.movement.zone_place
-        )
-    )
-
-    send_grip_release(client, CONF.tool.release_state)
-
-    client.send(
-        MoveToFrame(
-            bullet.placement_frame,
-            CONF.movement.speed_placing,
-            CONF.movement.zone_place,
-        )
-    )
-
-    client.send(
-        MoveToFrame(
-            offset_placement, CONF.movement.speed_travel, CONF.movement.zone_travel
-        )
-    )
-
-    # offset placement frame then safety frame
-    for frame in bullet.trajectory_from:
-        client.send(
-            MoveToFrame(frame, CONF.movement.speed_travel, CONF.movement.zone_travel)
-        )
-
-    client.send(StopWatch())
-    return client.send(ReadWatch())
-
-
-def send_grip_release(client, do_state):
-    """Grip or release using RCF tool, either in simulation or on real robot.
-
-    If script target is real robot this will set the digital output on the robot,
-    and if the target is virtual robot it will use RobotStudio code to visualize
-    clay bullet gripping and releasing
-
-    Parameters
-    ----------
-    client : :class:`compas_rrc.AbbClient`
-    do_state : int (0 or 1)
-        Value to set DO to
-    """
-    if CONF.target == "real":
-        client.send(WaitTime(CONF.tool.wait_before_io))
-        client.send(SetDigital(CONF.tool.io_needles_pin, do_state))
-        client.send(WaitTime(CONF.tool.wait_after_io))
-    else:
-        # Custom instruction can grip a bullet in RobotStudio
-        # note the tool tip must touch the bullet
-        if do_state == CONF.tool.grip_state:
-            client.send(CustomInstruction("r_A057_RS_ToolGrip"))
-        else:
-            client.send(CustomInstruction("r_A057_RS_ToolRelease"))
-
-    logging.debug(
-        "Signal sent to {}".format(
-            "grip" if do_state == CONF.tool.grip_state else "release"
-        )
-    )
-
-
-################################################################################
-# Non programs
-################################################################################
-
-
-def get_settings():
-    """Print and prompts user for changes to default configuration.
-
-    Parameters
-    ----------
-    target_select : str ('real' or 'virtual')
-        Target for script, either virtual robot controller or real. From argparse.
-    """
-    load_or_default = questionary.select(
-        "Load config or use default?", choices=["Default", "Load"], default="Default"
-    ).ask()
-
-    if load_or_default == "Load":
-        conf_file = ui.open_file_dialog(
-            initial_dir=DEFAULT_CONF_DIR, file_type=("YAML files", "*.yaml")
-        )
-        fabrication_conf.set_file(conf_file)
-        logging.info("Configuration loaded from {}".format(conf_file))
-    else:
-        fabrication_conf.read(defaults=True, user=False)
-        logging.info("Default configuration loaded from package")
-
-    if not fabrication_conf["target"].exists():
-        question = questionary.select(
-            "Target?", choices=["Virtual robot", "Real robot"], default="Virtual robot"
-        ).ask()
-        fabrication_conf["target"] = "real" if question == "Real robot" else "virtual"
-
-    logging.info(
-        "Target is {} controller.".format(fabrication_conf["target"].get().upper())
-    )
-
-    # At this point the conf is considered set, if changes needs to happen after
-    # this point CONF needs to be set again. There's probably a better way though.
-    global CONF
-    CONF = fabrication_conf.get(abb_rcf_conf_template)
-
-    logging.info("Configuration \n{}".format(fabrication_conf.dump()))
 
 
 def pick_frame_from_grid(index, bullet_height):
@@ -334,73 +58,147 @@ def pick_frame_from_grid(index, bullet_height):
     list of `class`:compas.geometry.Frame
     """
     # If index is larger than amount on picking plate, start from zero again
-    index = index % (CONF.pick.xnum * CONF.pick.ynum)
+    index = index % (fab_conf["pick"]["xnum"].get() * fab_conf["pick"]["ynum"].get())
 
-    xpos = index % CONF.pick.xnum
-    ypos = index // CONF.pick.xnum
+    xpos = index % fab_conf["pick"]["xnum"].get()
+    ypos = index // fab_conf["pick"]["xnum"].get()
 
-    x = CONF.pick.origin_grid.x + xpos * CONF.pick.grid_spacing
-    y = CONF.pick.origin_grid.y + ypos * CONF.pick.grid_spacing
-    z = bullet_height * CONF.pick.compression_height_factor
+    x = (
+        fab_conf["pick"]["origin_grid"]["x"].get()
+        + xpos * fab_conf["pick"]["grid_spacing"].get()
+    )
+    y = (
+        fab_conf["pick"]["origin_grid"]["y"].get()
+        + ypos * fab_conf["pick"]["grid_spacing"].get()
+    )
+    z = bullet_height * fab_conf["pick"]["compression_height_factor"].get()
 
-    frame = Frame(Point(x, y, z), Vector(*CONF.pick.xaxis), Vector(*CONF.pick.yaxis))
-    logging.debug("Picking frame {:03d}: {}".format(index, frame))
+    frame = Frame(
+        Point(x, y, z),
+        Vector(*fab_conf["pick"]["xaxis"].get()),
+        Vector(*fab_conf["pick"]["yaxis"].get()),
+    )
+    log.debug("Picking frame {:03d}: {}".format(index, frame))
     return frame
+
+
+def logging_setup():
+    timestamp_file = datetime.now().strftime("%Y%m%d-%H.%M_rcf_abb.log")
+    log_file = fab_conf["paths"]["log_dir"].get(Path()) / timestamp_file
+
+    handlers = []
+
+    if not fab_conf["skip_logfile"].get():
+        handlers.append(log.FileHandler(log_file, mode="a"))
+    if fab_conf["quiet"].get() is not True:
+        handlers.append(log.StreamHandler(sys.stdout))
+
+    log.basicConfig(
+        level=log.DEBUG if fab_conf["debug"].get() else log.INFO,
+        format="%(asctime)s:%(levelname)s:%(funcName)s:%(message)s",
+        handlers=handlers,
+    )
+
+
+def setup_fab_data(clay_bullets):
+    """Check for placed bullets in JSON.
+
+    Parameters
+    ----------
+    clay_bullets : list of :class:`compas_rcf.fabrication.clay_objs.ClayBullet`
+        Original list of ClayBullets.
+
+    Returns
+    -------
+    list of :class:`compas_rcf.fabrication.clay_objs.ClayBullet`
+        Curated list of ClayBullets
+    """
+    maybe_placed = [bullet for bullet in clay_bullets if bullet.placed is not None]
+
+    if len(maybe_placed) < 1:
+        return clay_bullets
+
+    last_placed = max(maybe_placed, key=attrgetter("bullet_id"))
+    last_placed_index = clay_bullets.index(last_placed)
+
+    log.info(
+        "Last bullet placed was {:03}/{:03} with id {}.".format(
+            last_placed_index, len(clay_bullets), last_placed.bullet_id
+        )
+    )
+
+    skip_options = questionary.select(
+        "Some or all bullet seems to have been placed already.",
+        [
+            "Skip all bullet marked as placed in JSON file.",
+            "Place all anyways.",
+            questionary.Separator(),
+            "Place some of the bullets.",
+        ],
+    ).ask()
+
+    if skip_options == "Skip all bullet marked as placed in JSON file.":
+        to_place = [bullet for bullet in clay_bullets if bullet not in maybe_placed]
+    if skip_options == "Place all anyways.":
+        to_place = clay_bullets[:]
+    if skip_options == "Place some of the bullets.":
+        skip_method = questionary.select(
+            "Select method:",
+            ["Place last N bullets again.", "Pick bullets to place again."],
+        ).ask()
+        if skip_method == "Place last N bullets again.":
+            n_place_again = questionary.text(
+                "Number of bullets from last to place again?",
+                "1",
+                lambda val: val.isdigit() and -1 < int(val) < last_placed_index,
+            ).ask()
+            to_place = clay_bullets[last_placed_index - int(n_place_again) + 1 :]
+            log.info(
+                "Placing last {} bullets again. First bullet will be id {}.".format(
+                    n_place_again, to_place[0].bullet_id,
+                )
+            )
+        else:
+            to_place_selection = questionary.checkbox(
+                "Select bullets:",
+                [
+                    "{:03} (id {}), marked placed: {}".format(
+                        i, bullet.bullet_id, bullet.placed is not None
+                    )
+                    for i, bullet in enumerate(clay_bullets)
+                ],
+            ).ask()
+            indices = [int(bullet.split()[0]) for bullet in to_place_selection]
+            to_place = [clay_bullets[i] for i in indices]
+
+    return to_place
 
 
 ################################################################################
 # Script runner                                                                #
 ################################################################################
-def abb_run(cmd_line_args):
+def abb_run():
     """Fabrication runner, sets conf, reads json input and runs fabrication process."""
-    fabrication_conf.set_args(cmd_line_args)
 
-    ############################################################################
-    # Logging setup                                                            #
-    ############################################################################
-    timestamp_file = datetime.now().strftime("%Y%m%d-%H.%M_rcf_abb.log")
-    log_file = DEFAULT_LOG_DIR / timestamp_file
-
-    handlers = []
-
-    if not fabrication_conf["skip_logfile"]:
-        handlers.append(logging.FileHandler(log_file, mode="a"))
-
-    if not fabrication_conf["quiet"]:
-        handlers.append(logging.StreamHandler(sys.stdout))
-
-    logging.basicConfig(
-        level=logging.DEBUG if fabrication_conf["debug"] else logging.INFO,
-        format="%(asctime)s:%(levelname)s:%(funcName)s:%(message)s",
-        handlers=handlers,
-    )
-
-    logging.info("compas_rcf version: {}".format(__version__))
-    logging.debug("argparse input: {}".format(cmd_line_args))
-    logging.debug("config after set_args: {}".format(fabrication_conf))
-
-    ############################################################################
-    # CONF setup                                                              #
-    ############################################################################
-    get_settings()
+    # CONF setup
+    interactive_conf_setup()
 
     ############################################################################
     # Docker setup                                                            #
     ############################################################################
     compose_up(docker_compose_paths["base"], remove_orphans=False)
-    logging.debug("Compose up base")
-    ip = robot_ips[CONF.target]
+    log.debug("Compose up base")
+    ip = robot_ips[fab_conf["target"].as_str()]
     compose_up(docker_compose_paths["abb_driver"], ROBOT_IP=ip)
-    logging.debug("Compose up abb_driver")
+    log.debug("Compose up abb_driver")
 
     ############################################################################
     # Load fabrication data                                                    #
     ############################################################################
-    json_path = Path(ui.open_file_dialog(initial_dir=DEFAULT_JSON_DIR))
-
+    json_path = pathlib.Path(ui.open_file_dialog(fab_conf["paths"]["json_dir"].get()))
     clay_bullets = load_bullets(json_path)
-    logging.info("Fabrication data read from: {}".format(json_path))
-    logging.info("{} items in clay_bullets.".format(len(clay_bullets)))
+    log.info("Fabrication data read from: {}".format(json_path))
+    log.info("{} items in clay_bullets.".format(len(clay_bullets)))
 
     ############################################################################
     # Create Ros Client                                                        #
@@ -412,27 +210,9 @@ def abb_run(cmd_line_args):
     ############################################################################
     abb = AbbClient(ros)
     abb.run()
-    logging.debug("Connected to ROS")
+    log.debug("Connected to ROS")
 
-    ############################################################################
-    # Connection check                                                         #
-    ############################################################################
-    ip = robot_ips[CONF.target]
-    for i in range(3):
-        try:
-            logging.debug("Pinging robot")
-            ping(abb, timeout=CONF.docker.timeout_ping)
-            logging.debug("Breaking loop after successful ping")
-            break
-        except TimeoutError:
-            logging.info("No response from controller, restarting abb-driver service.")
-            compose_up(
-                docker_compose_paths["abb_driver"], force_recreate=True, ROBOT_IP=ip
-            )
-            logging.debug("Compose up for abb_driver with robot-ip={}".format(ip))
-            time.sleep(CONF.docker.sleep_after_up)
-    else:
-        raise TimeoutError("Failed to connect to robot")
+    connection_check(abb)
 
     ############################################################################
     # setup in_progress JSON                                                   #
@@ -446,78 +226,15 @@ def abb_run(cmd_line_args):
             json_progress_identifier + json_path.name
         )
 
-    ############################################################################
-    # Check for placed bullets in JSON #
-    ############################################################################
-
-    maybe_placed = [bullet for bullet in clay_bullets if bullet.placed is not None]
-
-    if len(maybe_placed) < 1:
-        to_place = clay_bullets[:]
-    else:
-
-        last_placed = max(maybe_placed, key=attrgetter("bullet_id"))
-        last_placed_index = clay_bullets.index(last_placed)
-
-        logging.info(
-            "Last bullet placed was {:03}/{:03} with id {}.".format(
-                last_placed_index, len(clay_bullets), last_placed.bullet_id
-            )
-        )
-
-        skip_options = questionary.select(
-            "Some or all bullet seems to have been placed already.",
-            [
-                "Skip all bullet marked as placed in JSON file.",
-                "Place all anyways.",
-                questionary.Separator(),
-                "Place some of the bullets.",
-            ],
-        ).ask()
-
-        if skip_options == "Skip all bullet marked as placed in JSON file.":
-            to_place = [bullet for bullet in clay_bullets if bullet not in maybe_placed]
-        if skip_options == "Place all anyways.":
-            to_place = clay_bullets[:]
-        if skip_options == "Place some of the bullets.":
-            skip_method = questionary.select(
-                "Select method:",
-                ["Place last N bullets again.", "Pick bullets to place again."],
-            ).ask()
-            if skip_method == "Place last N bullets again.":
-                n_place_again = questionary.text(
-                    "Number of bullets from last to place again?",
-                    "1",
-                    lambda val: val.isdigit() and -1 < int(val) < last_placed_index,
-                ).ask()
-                to_place = clay_bullets[last_placed_index - int(n_place_again) + 1 :]
-                logging.info(
-                    "Placing last {} bullets again. First bullet will be id {}.".format(
-                        n_place_again, to_place[0].bullet_id,
-                    )
-                )
-            else:
-                to_place_selection = questionary.checkbox(
-                    "Select bullets:",
-                    [
-                        "{:03} (id {}), marked placed: {}".format(
-                            i, bullet.bullet_id, bullet.placed is not None
-                        )
-                        for i, bullet in enumerate(clay_bullets)
-                    ],
-                ).ask()
-                indices = [int(bullet.split()[0]) for bullet in to_place_selection]
-                to_place = [clay_bullets[i] for i in indices]
+    to_place = setup_fab_data(clay_bullets)
 
     if not questionary.confirm("Ready to start program?").ask():
-        logging.critical("Program exited because user didn't confirm start.")
+        log.critical("Program exited because user didn't confirm start.")
         print("Exiting.")
         sys.exit()
 
-    ############################################################################
-    # Set speed, accel, tool, wobj and move to start pos                       #
-    ############################################################################
-    initial_setup(abb)
+    # Set speed, accel, tool, wobj and move to start pos
+    pre_procedure(abb)
 
     for bullet in to_place:
         bullet.placed = None
@@ -533,23 +250,23 @@ def abb_run(cmd_line_args):
         )
 
         abb.send(PrintText(current_bullet_desc))
-        logging.info(current_bullet_desc)
+        log.info(current_bullet_desc)
 
         pick_frame = pick_frame_from_grid(i, bullet.height)
 
         # Pick bullet
-        pick_future = send_picking(abb, pick_frame)
+        pick_future = pick_bullet(abb, pick_frame)
 
         # Place bullet
-        place_future = send_placing(abb, bullet)
+        place_future = place_bullet(abb, bullet)
 
-        # This both sets cycle_time and blocks script until robot finishes pick & place
+        # This blocks until cycle is finished
         cycle_time = pick_future.result() + place_future.result()
 
         bullet.cycle_time = cycle_time
-        logging.debug("Cycle time was {}".format(bullet.cycle_time))
+        log.debug("Cycle time was {}".format(bullet.cycle_time))
         bullet.placed = time.time()
-        logging.debug("Time placed was {}".format(bullet.placed))
+        log.debug("Time placed was {}".format(bullet.placed))
 
         with in_progress_json.open(mode="w") as fp:
             json.dump(clay_bullets, fp, cls=ClayBulletEncoder)
@@ -559,28 +276,31 @@ def abb_run(cmd_line_args):
     ############################################################################
 
     if len([bullet for bullet in clay_bullets if bullet.placed is None]) == 0:
-        done_json = DEFAULT_JSON_DIR / "00_done" / json_path.name
+        done_file_name = json_path.name.replace(json_progress_identifier, "")
+        done_json = (
+            fab_conf["paths"]["json_dir"].get(Path()) / "00_done" / done_file_name
+        )
 
         in_progress_json.rename(done_json)
 
         with done_json.open(mode="w") as fp:
             json.dump(clay_bullets, fp, cls=ClayBulletEncoder)
 
-        logging.debug("Saved placed bullets to 00_Done.")
+        log.debug("Saved placed bullets to 00_Done.")
     else:
-        logging.debug(
+        log.debug(
             "Bullets without placed timestamp still present, keeping {}".format(
                 in_progress_json.name
             )
         )
 
-    logging.info("Finished program with {} bullets.".format(len(to_place)))
-    shutdown_procedure(abb)
+    log.info("Finished program with {} bullets.".format(len(to_place)))
+
+    post_procedure(abb)
 
 
 if __name__ == "__main__":
-    """Entry point and argument handling
-    """
+    """Entry point and argument handling."""
     import argparse
 
     parser = argparse.ArgumentParser()
@@ -605,4 +325,12 @@ if __name__ == "__main__":
 
     args = parser.parse_args()
 
-    abb_run(args)
+    fab_conf.set_args(args)
+
+    logging_setup()
+
+    log.info("compas_rcf version: {}".format(__version__))
+    log.debug("argparse input: {}".format(args))
+    log.debug("config after set_args: {}".format(fab_conf))
+
+    abb_run()