+++ tests/rcf/test_import.py
@@ -2,11 +2,6 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from compas_rcf.abb import *  # noqa E403
-from compas_rcf.fabrication import *  # noqa E403
-from compas_rcf.ur import *  # noqa E403
-from compas_rcf.utils import *  # noqa E403
-
 
 def test_nothing():
     assert 5 == 5